Adding auth checks for admin dashboard So that only staff can access and if not staff a custom view. Also shifted the top dashboard components to pages, as these are kind of what they are.
