--- src/components/EventDashboard/EventDashboard.css
@@ -1,11 +1,0 @@
-.dashboard {
-    display: flex;
-    flex-direction: row;
-    margin-top: 300px;
-    justify-content: space-between;
-    border: solid #134f5c;
-    border-radius: 20px;
-    margin-left:20px;
-    margin-right: 20px;
-    
-}
